Revert "Refactor event definition handling for search filter usage"
This reverts commit 0633495ff42127fa019afd0489b308a30b949678.
