+++ graylog2-server/src/main/java/org/graylog/events/processor/SearchFilterableConfig.java
@@ -21,7 +21,5 @@ import org.graylog.plugins.views.search.searchfilters.model.UsedSearchFilter;
 import java.util.List;
 
 public interface SearchFilterableConfig {
-    String FIELD_FILTERS = "filters";
-
     List<UsedSearchFilter> filters();
 }
